Splits the different items in their own class
Creates the _Item class which extends the original Item class, whilst adding the static handleQuality and handleSellin function. This class also implements the basic canQualityIncrease and isExpired (which do exactly what they are named after)

The four classes that are then created all extend the _Item and implement their own handleQuality and handleSellin functions (since they all have different implementations)

The updateQuality function is changed. It now returns a newly created list of items. These items are created depending on their name (instantiating the correct class object that were described previously). Once the object is correctly created, both its handleQuality and handleSellin func are called and it is then pushed to the returned list

Adds a conjuredMultiplier (determined from the name of the product)